requirements-server.txt
# Minimal requirements to run the FastAPI server
# (no torch / transformers / RL — keeps the image small)

gymnasium>=0.29.0
numpy>=1.24.0
pydantic>=2.0.0
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
rich>=13.0.0
python-dotenv>=1.0.0
openai>=1.0.0
httpx>=0.27.0
